make numba dependency optional
Allow building python 3.13 wheels without support for numba flow kernels
for now.
