+++ pyproject.toml
@@ -25,7 +25,6 @@ maintainers = [
 requires-python = ">=3.10"
 dependencies = [
     "numpy>=1.24",
-    "numba",
 ]
 
 [project.urls]
@@ -33,6 +32,7 @@ Home = "https://fastscapelib.readthedocs.io"
 Repository = "https://github.com/fastscape-lem/fastscapelib"
 
 [project.optional-dependencies]
+numba = ["numba"]
 test = ["pytest>=6.0"]
 
 [tool.scikit-build]
@@ -43,6 +43,12 @@ environment = {SKBUILD_CMAKE_ARGS='-DFS_DOWNLOAD_XTENSOR_PYTHON=ON'}
 test-extras = "test"
 test-command = "pytest {project}/python/fastscapelib/tests"
 build-verbosity = 1
+
+[[tool.cibuildwheel.overrides]]
+# TODO: remove when numba supports python 3.13
+select = "cp3{10,11,12}-*"
+inherit.test-requires = "append"
+test-requires = ["numba"]
 
 [tool.cibuildwheel.macos]
 environment = {SKBUILD_CMAKE_ARGS='-DFS_DOWNLOAD_XTENSOR_PYTHON=ON', MACOSX_DEPLOYMENT_TARGET=10.13}
